+++ sdk/storage/azure-storage-files-shares/src/share_service_client.cpp
@@ -111,7 +111,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
   ShareClient ServiceClient::GetShareClient(const std::string& shareName) const
   {
     auto builder = m_serviceUri;
-    builder.AppendPath(shareName);
+    builder.AppendPath(Storage::Details::UrlEncodePath(shareName));
     return ShareClient(builder, m_pipeline);
   }