+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.0-beta.3 (2020-10-13)
 
+### New Features
+
 * Service version is now 2020-02-10.
 * Added support for SAS generation in DataLake service.
 * CreateFromConnectionString now accepts unencoded file, path and directory name.
@@ -10,15 +12,26 @@
 * `ETag` and `LastModified` is now `std::string` instead of `Azure::Core::Nullable<std::string>` in `CreateDirectoryResult`, `CreateFileResult` and `CreatePathResult`.
 * `PathAppendDataResult` now returns `ContentMD5`, `ContentCrc64` and `IsServerEncrypted`.
 
+### Breaking Changes
+
+* `CreateFromConnectionString` now accepts unencoded file, path and directory name.
+* `ETag` and `LastModified` is now `std::string` instead of `Azure::Core::Nullable<std::string>` in `CreateDirectoryResult`, `CreateFileResult` and `CreatePathResult`.
+
 ## 1.0.0-beta.2 (2020-09-09)
+
+### New Features
 
 * Release based on azure-core_1.0.0-beta.1
 
 ## 1.0.0-beta.1
 
+### New Features
+
 * Support for Lease related operations.
 
 ## 1.0.0-preview.1 (Unreleased)
+
+### New Features
 
 * Added support for DataLake features:
   - ServiceClient::ListFileSystems